@@ -4,7 +4,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   equatable: ^2.0.3
@@ -16,4 +16,4 @@ dev_dependencies:
   build_runner: ^2.0.0
   json_serializable: ^6.0.0
   test: ^1.17.0
-  very_good_analysis: ^2.4.0
+  very_good_analysis: ^3.0.1
