@@ -4,13 +4,13 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.13.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   todos_api:
     path: ../todos_api
 
 dev_dependencies:
-  mocktail: ^0.2.0
+  mocktail: ^0.3.0
   test: ^1.17.0
-  very_good_analysis: ^2.4.0
+  very_good_analysis: ^3.0.1
