@@ -4,7 +4,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   bloc: ^8.0.1
@@ -28,9 +28,9 @@ dev_dependencies:
   bloc_test: ^9.0.1
   flutter_test:
     sdk: flutter
-  mockingjay: ^0.2.0
-  mocktail: ^0.2.0
-  very_good_analysis: ^2.4.0
+  mockingjay: ^0.3.0
+  mocktail: ^0.3.0
+  very_good_analysis: ^3.0.1
 
 flutter:
   uses-material-design: true
