@@ -35,7 +35,9 @@ class LocalStorageTodosApi extends TodosApi {
   void _init() {
     final todosJson = _getValue(kTodosCollectionKey);
     if (todosJson != null) {
-      final todos = List<Map>.from(json.decode(todosJson) as List)
+      final todos = List<Map<dynamic, dynamic>>.from(
+        json.decode(todosJson) as List,
+      )
           .map((jsonMap) => Todo.fromJson(Map<String, dynamic>.from(jsonMap)))
           .toList();
       _todoStreamController.add(todos);
